+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix using bytes file paths for `CDNReader` and `SDFilesystem`
 * Support auto-closing underlying file for `CTRFileIO` and `CBCFileIO`
 * Make `CTRFileIO` and `CBCFileIO` objects hashable (if the underlying file object is)
+* Update `CDNReader` to re-open files instead of using shared file objects, and internally open all files through `CDNReader.open_raw_section` (fixes #6)
 
 ## v0.4.5 - October 24, 2020
 * Fix loading RomFS from a filename in `RomFSReader`